chore: move managed udf tests to large tests
Right now we need to actively cleanup the managed BQ functions
b/403599621, so we are moving away from creating persistent managed functions.
